+++ ASP.NET-MVC/BookShop/BookShop.AcceptanceTests/Drivers/Integrated/IntegratedBookDetailsDriver.cs
@@ -1,5 +1,6 @@
 ﻿using System.Linq;
 using BookShop.AcceptanceTests.Support;
+using BookShop.AcceptanceTests.Support.Database;
 using BookShop.Mvc.Controllers;
 using BookShop.Mvc.Logic;
 using BookShop.Mvc.Models;
