+++ ASP.NET-MVC/BookShop/BookShop.AcceptanceTests/Drivers/Integrated/IntegratedHomeDriver.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using BookShop.AcceptanceTests.Common;
 using BookShop.AcceptanceTests.Support;
 using BookShop.Mvc.Controllers;
 using BookShop.Mvc.Logic;
