+++ ASP.NET-MVC/BookShop/BookShop.AcceptanceTests/Drivers/Integrated/IntegratedShoppingCartDriver.cs
@@ -1,6 +1,8 @@
 ﻿using System;
 using System.Linq;
 using BookShop.AcceptanceTests.Support;
+using BookShop.AcceptanceTests.Support.Database;
+using BookShop.AcceptanceTests.Support.Webserver;
 using BookShop.Mvc.Controllers;
 using BookShop.Mvc.Logic;
 using BookShop.Mvc.Models;
