+++ ASP.NET-MVC/BookShop/BookShop.AcceptanceTests/Support/Database/CatalogContext.cs
@@ -1,4 +1,4 @@
-﻿namespace BookShop.AcceptanceTests.Support
+﻿namespace BookShop.AcceptanceTests.Support.Database
 {
     public class CatalogContext
     {
