+++ ASP.NET-MVC/BookShop/BookShop.Mvc/BookShop.Mvc.csproj
@@ -14,7 +14,7 @@
       <PrivateAssets>all</PrivateAssets>
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
     </PackageReference>
-    <PackageReference Include="Moq" Version="4.14.1" />
+    
     <PackageReference Include="Newtonsoft.Json" Version="12.0.3" />
   </ItemGroup>
